+++ src/main/java/net/fracturedcode/xpbeacons/XpBeaconsExtension.java
@@ -3,6 +3,8 @@ package net.fracturedcode.xpbeacons;
 import carpet.CarpetExtension;
 import carpet.CarpetServer;
 import carpet.settings.SettingsManager;
+import net.fracturedcode.xpbeacons.XpBeaconsCategorySettings.BeaconSettings;
+import net.fracturedcode.xpbeacons.XpBeaconsCategorySettings.EffectSettings.*;
 
 public class XpBeaconsExtension implements CarpetExtension {
     private final static SettingsManager customSettingsManager;
@@ -16,14 +18,14 @@ public class XpBeaconsExtension implements CarpetExtension {
     public void onGameStarted() {
         CarpetServer.settingsManager.parseSettingsClass(XpBeaconsSimpleSettings.class);
 
-        customSettingsManager.parseSettingsClass(XpBeaconsCategorySettings.HasteSettings.class);
-        customSettingsManager.parseSettingsClass(XpBeaconsCategorySettings.SpeedSettings.class);
-        customSettingsManager.parseSettingsClass(XpBeaconsCategorySettings.RegenerationSettings.class);
-        customSettingsManager.parseSettingsClass(XpBeaconsCategorySettings.ResistanceSettings.class);
-        customSettingsManager.parseSettingsClass(XpBeaconsCategorySettings.JumpBoostSettings.class);
-        customSettingsManager.parseSettingsClass(XpBeaconsCategorySettings.StrengthSettings.class);
+        customSettingsManager.parseSettingsClass(HasteSettings.class);
+        customSettingsManager.parseSettingsClass(SpeedSettings.class);
+        customSettingsManager.parseSettingsClass(RegenerationSettings.class);
+        customSettingsManager.parseSettingsClass(ResistanceSettings.class);
+        customSettingsManager.parseSettingsClass(JumpBoostSettings.class);
+        customSettingsManager.parseSettingsClass(StrengthSettings.class);
 
-        customSettingsManager.parseSettingsClass(XpBeaconsCategorySettings.BeaconSettings.class);
+        customSettingsManager.parseSettingsClass(BeaconSettings.class);
     }
 
     @Override
